Create info.py
## Extended Description

This Python file stores basic information such as a person's name, age, and height. 
It then prints these values to the console. 

The script is simple and is meant for beginners learning how to define variables 
and output text using the `print()` function in Python.

### What the program does:
- Stores a name as a string
- Stores age as a number
- Stores height in meters
- Prints all the information in a readable format

### Example Output:
Name: Peter  
Age: 9  
Height: 1.46 m